rename assets to public
